src/digi_ad/requirements.txt
# Core requirements
fastapi==0.115.9
uvicorn==0.34.2
pydantic==2.11.5
python-dotenv==1.1.0

# Database/Storage
sqlalchemy==2.0.30
chromadb==0.5.23
lancedb==0.21.1
qdrant-client==1.14.2

# AI/ML
openai==1.75.0
cohere==5.15.0
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.61
crewai==0.118.0
tiktoken==0.9.0
huggingface-hub==0.30.2

# Utilities
requests==2.32.3
httpx==0.28.1
orjson==3.10.18
rich==13.9.4
tenacity==9.1.2
watchfiles==1.0.5
